test: Split tests into conventionally named files
Replace test_new_features.py with three files following project conventions:
- test_diff_filtering.py — Diff.filter() and Diff.exclude()
- test_diffsync_diff_and_sync_parameters.py — model_types, filters,
  sync_attrs, exclude_attrs, sync_filter, concurrent, and combinations
- test_diffsync_model_bulk.py — bulk CRUD and store bulk operations

Converted from class-based to flat function style with Apache license
headers, matching the existing test structure.
